Implemented basic transaction editing

+++ transaction.go
@@ -3,14 +3,14 @@ package main
 import "time"
 
 type Transaction struct {
-	ID           int
+	ID           int64
 	Date         time.Time
 	Payee        string
 	Memo         string
-	Debit        int
-	Credit       int
-	CategoryID   int
-	AccountID    int
+	Debit        int64
+	Credit       int64
+	CategoryID   int64
+	AccountID    int64
 	IsCleared    bool
 	IsReconciled bool
 }
